+++ http4k-contract/src/test/kotlin/org/http4k/contract/RouteTest.kt
@@ -4,6 +4,7 @@ import com.natpryce.hamkrest.absent
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import com.natpryce.hamkrest.present
+import com.natpryce.hamkrest.should.shouldMatch
 import com.natpryce.hamkrest.throws
 import org.http4k.core.Body
 import org.http4k.core.ContentType.Companion.TEXT_PLAIN
@@ -12,11 +13,13 @@ import org.http4k.core.Method.GET
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
+import org.http4k.core.Uri
 import org.http4k.core.with
 import org.http4k.lens.Header
 import org.http4k.lens.LensFailure
 import org.http4k.lens.Path
 import org.http4k.lens.Query
+import org.http4k.lens.int
 import org.http4k.lens.missing
 import org.http4k.lens.string
 import org.junit.Test
@@ -44,6 +47,18 @@ class RouteTest {
         assertThat((route.router(Root))(invalidRequest), present())
         assertThat({ (route.router(Root))(invalidRequest)?.invoke(invalidRequest) },
             throws(equalTo(LensFailure(query.meta.missing()))))
+    }
+
+    @Test
+    fun `can build a request from a route`() {
+        val path1 = Path.int().of("sue")
+        val path2 = Path.string().of("bob")
+        val route = Route("").at(GET) / path1 / path2
+        val request = route.newRequest(Uri.of("http://rita.com"))
+
+        request.with(path1 of 123, path2 of "hello world") shouldMatch equalTo(
+            Request(GET, "http://rita.com/123/hello+world")
+        )
     }
 
     @Test