+++ http4k-core/src/main/kotlin/org/http4k/lens/Path.kt
@@ -1,7 +1,6 @@
 package org.http4k.lens
 
 import org.http4k.core.Request
-import org.http4k.core.Uri
 import org.http4k.core.decode
 import org.http4k.core.encode
 import org.http4k.lens.ParamMeta.BooleanParam
@@ -77,7 +76,7 @@ open class BiDiPathLensSpec<OUT>(paramMeta: ParamMeta,
 
 object Path : BiDiPathLensSpec<String>(StringParam,
     LensGet { _, target -> listOf(target.decode()) },
-    LensSet { name, values, target -> target.uri(Uri.of(target.uri.path.replaceFirst("{$name}", values.first().encode()))) }) {
+    LensSet { name, values, target -> target.uri(target.uri.path(target.uri.path.replaceFirst("{$name}", values.first().encode()))) }) {
 
     fun fixed(name: String): PathLens<String> {
         val getLens = get(name)