+++ src/test/kotlin/cookbook/typesafe_http_requests_with_lenses.kt
@@ -15,20 +15,20 @@ import org.http4k.lens.int
 import org.http4k.lens.string
 
 /**
- * Example showing how to create and apply lenses to requests and responses. Not
+ * Example showing how to create and apply lenses to requests and responses.
  */
 fun main(args: Array<String>) {
 
     data class Child(val name: String)
 
-    val nameQuery = Header.required("name")
+    val nameHeader = Header.required("name")
     val ageQuery = Query.int().optional("age")
     val childrenBody = Body.string(TEXT_PLAIN).map({ it.split(",").map(::Child) }, { it.map { it.name }.joinToString() }).toLens()
 
     val endpoint = {
         request: Request ->
 
-        val name: String = nameQuery(request)
+        val name: String = nameHeader(request)
         val age: Int? = ageQuery(request)
         val children: List<Child> = childrenBody(request)
 
@@ -41,11 +41,16 @@ fun main(args: Array<String>) {
 
     val app = ServerFilters.CatchLensFailure.then(endpoint)
 
-    val goodRequest = Request(Method.GET, "http://localhost:9000").header("name", "Jane Doe").query("age", "25").body("rita,sue,bob")
+    val goodRequest = Request(Method.GET, "http://localhost:9000").with(
+        nameHeader of "Jane Doe",
+        ageQuery of 25,
+        childrenBody of listOf(Child("Rita"), Child("Sue")))
 
     println(listOf("", "Request:", goodRequest, app(goodRequest)).joinToString("\n"))
 
-    val badRequest = Request(Method.GET, "http://localhost:9000").header("name", "John Doe").query("age", "too old!")
+    val badRequest = Request(Method.GET, "http://localhost:9000")
+        .with(nameHeader of "Jane Doe")
+        .query("age", "some illegal age!")
 
     println(listOf("", "Request:", badRequest, app(badRequest)).joinToString("\n"))
 }